Remove squeeze force from VTK OpenXR Actions
"squeezeforce" action is not implemented which leads to a warning.
Squeeze force is supported by the Valve Index (cf. OpenXR specs §6.4.9)
Because squeeze force is not a basic VTK action it can't have a generic
implementation. User can add this action by itself using
vtkOpenXRRenderWindowInteractor::AddAction and adding the corresponding
mapping in the JSon files.