Use the static loader when generating .pyi files
The generate_pyi.py script needs to import the modules that it
is generating .pyi files for.  For static builds, this means
that it must import the special "static importer" module.

Note that VTK's own "static importer" module is already loaded
by vtkmodules/__init__.py, but similar mechanisms are not always
present (or even desired) for other wrapped packages, e.g.
Paraview's.