vtkSOADataArrayTemplate: fixing GetVoidPointer()
Previously GetVoidPointer() should have been used as read only
but there is nothing preventing developers from using the returned
void pointer to try and modify the vtkDataArray's data. This turns
into a bug since modifying data in the void pointer array doesn't
guarantee to modify data in the vtkDataArray. This change allows
the vtkSOADataArray to work as expected as a vtkDataArray.