vtkOpenGLProjectedTetrahedraMapper: fix depth buffer blit
Fix depth buffer copy when a floating point framebuffer is available.
When the bit depth of VTK's render FBO and the one used by this
mapper are not the same, direct blitting from one FBO to the
other does not work. Fix that by separating the color buffer blit
from the depth buffer blit.