Use vtkIdTypeArray::Superclass over just vtkIdTypeArray
The super-class of `vtkIdTypeArray` is more general, and yet supports almost
all the functionality of the more derived type. Using it makes the filter work
in more cases.

Foe example, this is needed to work correclty when the input data is created by
a vtk-m based filter upstream (vtk-m mostly works with the generic array types).