vtkCamera: Various improvements to head tracking in CAVE
The view rendered in the CAVE was coupled to the desktop camera,
requiring CAVE users to move the desktop camera to the origin,
and then not touch it again, or it would affect the view seen
in the CAVE views.  This change decouples the camera transform
from the view transform when in off-axis mode to fix that problem.

The lighting of objects was broken in the CAVE.  In the best
case, the light seemed to come from the opposite side it should
have (try turning off the light kit and adding a single directional
light before this change).  By placing the eye offset into the
view transform in off-axis mode, objects are now lit on the
correct side.

Clipping was broken in the CAVE so that sometimes rendered objects
would disappear from one CAVE screen before the others.  The
problem was worse when viewing CAVE displays from more oblique
angles.  This changes fixes that problem by adding (or subtracting)
the size of the screen in world coordinates to the far (or near)
clipping planes.

GPU volume rendering under CAVE head tracking was broken as well,
as the camera position (and not the eye position) was used as
the starting point for tracing rays.  This change fixes that by
using the eye point instead, when in off-axis mode.