vtkEnSightWriter: support polyhedra, fix broken elem. types
- Add support for VTK_POLYHEDRON cells.
- Fix malformed output for VTK_POLYGON and VTK_CONVEX_POINT_SET cells.
- Fix node ordering for VTK_WEDGE, VTK_QUADRATIC_WEDGE and VTK_QUADRATIC_EDGE
  cells (VTK and EnSight use different implicit ordering for these types).

This change brings vtkEnSightWriter to parity with vtkEnSightGoldBinaryReader
with regards to supported element types: one can read a vtkUnstructuredDataSet
from EnSight Gold case, save it, and read it again without crashing or losing
cells. Previously this was problematic due to poor support for cells with
variable number of nodes (VTK_POLYHEDRON was being skipped and VTK_POLYGON
created malformed files).

The issue with previous VTK_POLYGON and VTK_CONVEX_POINT_SET implementation
was that it did not handle "nsided" and "nfaced" connectivity representation
which is different from other cells that have constant number of nodes.
Please see comments in code for more details.