vtkVRRenderer: Fix hard coded translation when resetting camera
In VR, resetting the camera would place the object at floor level. For
this reason, a translation was added to bring the object closer to eye level.
However, this translation only worked if the view up direction (the
direction of the top of the head) was +Y.
This commit takes into account the current view up direction to always
bring the object to eye level.