Add VTK_PYTHON_FULL_THREADSAFE as option
The VTK_PYTHON_FULL_THREADSAFE variable has existed since 2015, but
so far has only been a cmake option in Paraview, never in VTK itself.
Since this option adds the GIL locks that are required for any
Python concurrency via PyEval_SaveThread()/PyEval_RestoreThread(),
it should be useful within a broad range of VTK projects.

The VTK_NO_PYTHON_THREADS option is an older option that was available
in VTK 5 but disappeared in VTK 6, though it still appeared in the C++
source files and in Paraview's CMakeLists.txt.