Move check for visibility to leaf translucency check
When deciding if a composite dataset has translucency,
vtkCompositePolyDataMapper2 was incorrectly determining there was no
translucency for a subtree of blocks if the visibility of the subtree
root was set to false. This is incorrect because children could be set
to visible, but the algorithm would not descend the tree to determine
that.

Fixed this bug by moving the visibility determination down to just
before the scalars translucency determination is made. If the block is
not visibile, it does not have translucency.