Adjust VTK_AXIS_EPSILON for GetOrientation
The vtkTransform::GetOrientation() method computes the orientation
angles from the transform's matrix elements.  For an ill-conditioned
matrix the results are not accurate, and in order to avoid giving
nonsense results, the code gives angles of 0 for condition numbers
greater than 1000.  The condition number is the factor by which
errors in the inputs (the matrix elements) are magnified in the
outputs (the angles).

Originally the matrix used single-precision floats, where the
numerical errors on the elements could be considered to be roughly
1e-6 (a little larger than machine epsilon).  So a condition number
of 1000 would give an error in the angles of around 1e-3.

Now that the matrix uses double-precision floats (since VTK 4.4) the
the error elements is likely to be much smaller, probably around
2e-15 (again, somewhat larger than machine epsilon).  So we can
tolerate a far larger condition number, as large as 500000000000.