Fix grounded movement with OpenXR
Fix the "Grounded" movement bug, that was not working with OpenXR selected as backend.
The reason was the HeadSetDir variable, that is needed for movement computations,
was never set in the vtkVRInteractorStyle with OpenXR.

This simply get the HMD pose (for now left eye) from the runtime and
emits a new event in vtkOpenXRRenderWindowInteractor::PollXrActions.