Change search path for vtkRenderingTk library
The best place to start is the vtkmodules dir itself, in case the
DLLs are installed in the same dir as the extension modules, and
the next best place is the "lib" directory two directories up from
the vtkmodules directory.

It would also be possible to generate the exact location of the
library with cmake, but we'll try a more general search first.