vtkThreadedCallbackQueue: changing SetNumberOfThreads behavior
`vtkThreadedCallbackQueue::SetNumberOfThreads` used to stop all running
threads, then allocate new threads but not run them. The user would have
had to restart the threads. As a consequence, we go from a state of `n`
threads to a state of zero threads to a state of `m` threads. Ideally,
we'd like the transition to be smoother so we just directly go from `n`
to `m` threads.

The method has been updated to behave like the latter scenario.
`SerialStop` is not needed anymore either so it is removed.