Add rendering for the `vtkCellGrid` data object.
Note that unlike other rendering (which always extracts
surfaces and renders those), the `vtkCellGridMapper` and
its OpenGL implementation send the entire mesh to the
GPU and use shaders to render cell surfaces.

This is a step toward in-situ simulation and visualization
as well as interactive editing using shaders to manipulate
arrays.

Co-Authored-By: Jaswant Panchumarti <jaswant.panchumarti@kitware.com>
Co-Authored-By: David Thompson <david.thompson@kitware.com>