Fix Python overrides that use existing name
The 'override' function in the wrappers has a check to ensure that
users don't override one wrapped C++ class with another wrapped
C++ class (see a4b954b8).  The overriding class must be a Python
class that is a subclass of the overridden C++ class.

However, rtImageTest.py defined a class called vtkTestingInteractor
and an override with this class failed because within VTK there is
a class called vtkmodules.vtkTestingRendering.vtkTestingInteractor.
The check has to consider not only the name, but also the type, or
else it will get false positives.