Remove NO_RT marker from tests that don't need it
The only tests that require NO_RT are the ones that require "-B"
to get the baseline directory instead of "-V" to get the actual
baseline image.  In other words, the tests that require NO_RT
are the ones that use multiple baseline images, or that use a
baseline image that isn't named after the test.