vtkThreadedCallbackQueue: optimizing Wait a tiny bit.
This commit makes that `vtkThreadedCallbackQueue::Wait` can run tasks
that are hanging if they are not on hold. This should save time on a
jammed queue.

A few extra boolean states needed added to InvokerFutureSharedStateBase.
We can actually merge them all on a binary mask. Each state is
exclusive.

Finally, a method `vtkThreadedCallbackQueue::Invoke` has been introduced
to avoid bugs in the future. One should invoke an invoker by passing it
to this function, which makes sure that the invoker is invoked in the
correct way.