vtkThreadedCallbackQueue: removing the Controller
It turns out that the queue was already almost capable of performing
asynchronous controls without an external controller. We're saving a
thread by doing so.

The asynchronous controls are now executed by the queue itself. We
ensure that they execute one after another by storing the futures of the
controls now done yet in a container called `ControlFutures`. Controls
have a high priority of execution and are placed in the front of the
queue rather than in the back. When a control is done doing its stuff,
it removes its future from `ControlFutures`.

There's not really any point to `Start()` anymore, so this private
function is removed.