vtkThreadedCallbackQueue: merging invoker and future
The class hierarchy of invokers and futures is changed. Now `vtkInvoker`
inherits from `vtkSharedFuture`. This allows to reduce the number of
allocation when pushing a task. There used to be 3 needed, now one
allocation is enough. This also allows to get rid of `InvokersOnHold`,
as the invoker is directly accessible from `Dependents`. There is no
more need to fetch the invoker associated with a future. You can just
run it when the invoker is ready.