ENH: Support string types in ArrayListTemplates
This change adds template specialization of `ArrayPair` struct for
`vtkStdString` type.

This was required so as to allow point and cell string arrays to be
copied through to output in vtkAlgorithm classes.

This change fixes the issue raised where catagorical colors were not
working in ParaView as it relies on vtkGeometryFilter to pass the
dataset attributes from input to output.