Add 3D Cursor Widget
Add a new widget representing a 3D cursor. This cursor is intended to be used in a stereo environment.
It takes form of an actor (a cross by default) that is added to the scene, hence it's displayed in stereo like other actors.
We take advantage of the existing API for widget handles in vtkAbstractWidget to insure the cursor keeps the same size on the display.