Remove vtkDataObject dependency from CommonCore
By assuming that the address of the object does not change when
upcasting and downcasting, as long as there is single inheritance
between the types, we can eliminate the use of vtkDataObject.h in
CommonCore.  This removes a hidden private dependency cycle
between CommonCore and CommonDataModel that could sometimes be
seen as undefined symbols.