Fix MetaIO ElementSpacing precision mismatch
Between the VTK 8 release and the VTK 9 release, the MetaIO library
changed the ElementSpacing type from float to double.  However, the
vtkMetaImageWriter was not changed to match.  Instead, it rounded
the vtkImageData spacing from double to float, and then the MetaIO
API converted it back to double before printing the file header.
Also note that the MetaIO file header changed from using 6 decimal
digits of precision for spacing in VTK 8, to 17 decimal digits of
precision for VTK 9.