vtkConvertSelection: Change block indices to selectors' compositeIndices
We get the composite ids from the selectors that correspond to the indices.
This is done to avoid selecting pieces/datasets from a partitioned/multi-piece dataset
and selecting only partitioned/multi-piece datasets, except if the parent of the index is
a multiblock.

This way block_selectors-selections and block-selections now generate the same result.