Add neighboring point search modes
Add an enum to configure how the filter selects the neighbor points used to
calculate the PCA. 2 modes available:
 - KNN: By default, K points are selected regardless of their location
relative to the sampled point. The radius can also be set to ensure that a
sufficiently large neighborhood is taken into account, if not the next
approach is performed.
 - RADIUS: A second approach is to select neighboring points inside a radius,
only the neighborhood of the sampled is considered. If K is also set, the
number of points found inside the radius must be larger than K, if not the
first approach is performed.

Both approaches give the same results. The first approach is faster for uniform
point clouds, in other cases, the second approach is faster.

Initialize attribute in the header file.

Update TraverseAndFlip to be consistent with what is done in the
estimation of normals. Refactor in a FindPoints function of the nearest
neighbor search behavior.