Add vtkPolyhedronUtilities and Decompose method
Add a new method allowing to decompose polyhedron into tets.
This method will generate new points on polyhedron and it's purpose
is to improve the result of subsequent filters (e.g. contours) on
polyhedrons with concave faces.

The vtkPolyhedronUtilities class is added to contain this method, and
potential future ones working on polyhedrons for the same kind of purpose.
This class can be seen as a toolbox for vtkPolyhedron.