IO/OCCT: drop components and check for targets
OpenCASCADE's component handling leaves a lot to be desired. First, it
requires the caller sort the list of components which also means that
interleaving with `OPTIONAL_COMPONENTS` is not possible. Second, the
components are more like VTK's "groups" and don't actually correspond to
full sets of targets. Instead, just find OpenCASCADE and check for
targets manually.