vtkEnSightWriter: write .case file in RequestData
With this change, Paraview writes .case file as well as data files
when saving to EnSight Gold. Previously the .case file was missing.

When running in parallel we want to write .sos file as well - this is the
master file that links individual .case pieces written by processes
into one logical case.