Wrap three more variants of vtkCellData::SetData.
The only variant of vtkCellData::SetData() (which took
vtkDataArrays) was not working well because of its implementation.
It makes shallow copies of its arguments even when they are
of the right type which causes issues with memory management
when the data comes from numpy (because vtkDataArrayPython is
the object that holds a reference to the numpy object not the
underlying buffer object). So added three more signatures which
directly take the right array types needed for things to work.