vtkTableBasedClipDataSet: Simplify using implicit structures
This simplification is possible thanks the addition of vtkDataSet::GetPoints()
and the addition of implicit points, cells and cell types structures for
structured data.

GetTuple creates a copy of the tuple using GetTypedTuple if it's not a vktDataArray
we do that since the input points can be implicit points, and GetTypedTuple is faster
than accessing the component of the TupleReference using GetTypedComponent internally.

Finally, simplify EvaluatePoints function by having only one functor and using the
vtkImplicitFunction::FunctionValue function which has been multithreaded.