Fix Java build error for templated parameters
The Java wrappers were not blocking templated parameters, even
though they are unable to wrap them.  This resulted in build
errors for methods like this one:

void AddArray(vtkAOSDataArrayTemplate<int>* array);

The fix for these build errors is to reject all methods with
parameter types that include "<" as part of their name.