Fixed buffer over-read in FiltersCorePython-TestBinnedDecimation
Reversed order of tests in for loop condition to check *first* that curPt is in-range before dereferencing it.

Also fixed off-by-1 in check for the end of the buffer.

Also fixed a UB I just noticed reading this code: endBatchPt was initialized to something potentially beyond the end of the buffer. Even though it was never dereferenced, even creating such a pointer is technically UB.