Fixed std::vector container-overflow found by ASan
The bug here was that even though `w.reserve()` was called, nothing was actually *added* to the `w` vector.  The later use of `w.data()` is thus not guaranteed to return a non-null pointer, as data() is documented to possibly return null for empty vectors.

As only use of `w` is as `double*` (it's stored into a `vtkSPDF_InternalParams` struct which uses `double*`), just got rid of the std::vector indirection.

Occurred during FiltersCorePython-smoothMeshOnMesh test.