vtkPolarAxesActor: Improve radial axes and ticks handling
Allow choosing a specific angle for radial axes.
This will either fill with as much radial axes it can, or requested count.
If set to 0, will find an ideal step depending on requested axes count.
This means both count and delta angle can't be set to 0 (auto).
Add an option for ticks to use radial axes angle (default to true).