vtkPolarAxesActor: Improve polar axes and ticks handling
Allow choosing a specific range for polar axes.
This will either fill with as much polar axes it can, or requested count.
If set to 0, will compute it as equal step depending on requested axes count.
This means both count and delta angle can't be set to 0 (auto).
Add an option for ticks to use radial axes angle (default to true).