Fix a bug in `tovtkm::ConvertSingleType`
There was a bug, where the worklet `ReorderHex` was being explicitly marked as
device only (not using VTKM_EXEC) when CUDA was enabled. However, in
`BuildSingleTypeVoxelCellSetVisitor` where the worklet is invoked, the
Cuda device was explicitly being disabled. This results in undefined
behaviour. This path is only trigered when converting an unstructured
grid containing all Voxels to vtkm DataSet.