Fix vtkXOpenGLRenderWindow hang in SetSize
Previously an XIfEvent was used to wait until a ConfigureNotify
indicated that the resize had occurred, but this would sometimes
hang the program indefinitely.  In particular, this could occur
sporadically when using vtkQtRenderWindowInteractor.py due to
race conditions beween Qt's and VTK's event handling.  But it
could hypothetically even occur simply due to a lack of response
from the window manager.  A delay loop with a 2s timeout is a
safer way to wait for the ConfigureNotify event.