vtkTableBasedClipDataSet: Save case index
The CellsMap array which stored if a cell generates anything
has been converted to a CellsCase array to store the cells' case index.
This avoids re-computation of the case index in the extract cells pass of the
algorithm which requires access to the clipArray. Therefore, we
have better memory access and computational performance.