Add ability for PIO Reader to load hdf5 variant of pio dumps
There is now a new variant of pio dumps, which are written in hdf5, but
still follow the pio format. The PIO Reader can now read pio dumps
written in this hdf5 variant. There is no difference in the resulting
data between loading a pio dump in native format versus loading the hdf5
variant. Added a new test to test this variant. Added copyright
information for the PIO Reader.