vtkFloatingPointExceptions: use `extern "C"` for `signal_handler`
C++17 allows non-C functions to be put here. However, this handler is
already very suspect because it uses `cerr` which is generally very not
allowed in signal handlers (there's a lock involved somewhere and a
deadlock is going to happen if another thread is using `cerr`).

Anyways, just resolve the `bugprone-signal-handler` lint for now.