vtkIOSSReader/Writer: MPI support fixes
As of now netcdf mpi support is not working for IOSSReader/Writer
because mpi calls are called inside the reader/writer instead of the ioss library
so we are using comm_null(), instead of comm_world().
In the future, when using comm_world() is used when and SEACAS_HAVE_MPI is on
my_processor and processor_count properties should be removed for exodus.
For more info. see Ioex::DatabaseIO::DatabaseIO in the ioss library.