docs: move build settings to its own document
This way building document becomes less intimidating and easier to read.

Also, a separate page exposes better this nice content instead of having it as
a section of building. It also allows us to create a dedicated page on
readthedocs more easily.

Lastly, it removes the unused NvPipe link originally introduced in
9491fe1a6 (docs: add documentation for building VTK). Considering he
NvPipe is deprecated since 2020, referencing it would not be relevant.

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>