vtkStructuredDataPlaneCutter: Remove ghosts after FlyingEdges if possible
vtkFlyingEdgesPlaneCutter does not remove ghosts. If it interpolated its
cell attributes and ghost are present, then we can remove them using RemoveGhostCells.
This way, we use the highly optimized vtkFlyingEdgesPlaneCutter instead of
vtkStructuredDataPlaneCutter which is also fast, but not as fast.

Also improve the speed of checking if ghost/blank cells are present using a bit mask.
Finally, also check if refined cells are present to produce the correct result for AMR.
That's why the number of output cells in the TestCompositeCutter needs to change.