ImplicitArray: NewInstance to give explicit AOS arrays
Although coutnerintuitive, changing the `NewInstance` behavior of
implicit arrays corresponds to more use cases then their current
behavior. The most common use of `NewInstance` in VTK is to create an
empty instance of an object and populate it with some filtered data. In
the case of implicit arrays, this population step is not possible  since
they are read only. Return AOS arrays allows most use cases to just work
out of the box.