Support any texture with interpolate scalars
VTK !8186 added support for using multitexturing with
InterpolateScalarsBeforeMappingOn to allow for blending interpolated
scalars with texture coloring.
This commit extends this support to the "actortexture" and "albedoTex"
textures that can be respectively set with vtkActor::SetTexture() and
vtkActor::GetProperty()->SetBaseColorTexture().

When turning on InterpolateScalarsBeforeMapping, the mapper internally
creates a texture with the associated texture coordinates.
This commit adds a specific attribute for those tcoords to avoid overriding
the existing polydata tcoords.