Support image direction in vtkImageResliceMapper
These changes allow vtkImageResliceMapper to give correct results
for oriented input images (that is, when the Direction isn't the
identity matrix).  The vtkImageSliceMapper already supported these
oriented images.

At the same time, the mapper was reformulated so that it now uses
the new vtkImageReslice::SetOutputDirection() method instead of
using vtkImageReslice::SetResliceAxes().