Fix reslice mapper depth issue for Alpha test
Adjust the reslice output to put it into a coordinate system that is
aligned with the SlicePlane.  Since this coordinate system depends
only on the SlicePlane and nothing else, it is identical for all
instances of vtkImageResliceMapper that use the same slice plane,
regardless of whether the prop orientation or image orientation
differs between mappers.  Hence we expect the mappers to paint their
pixels at the same depth in the depth buffer, reducing z-fighting.

This fixes the test regression for TestResliceMapperAlpha, which
failed due to z-fighting (overlay image wasn't fully rendered due
its depth values were sometimes greater than the underlay image).