revert the autotonemapping applied to ospray pathtracer
The effectively reverts 2449a38d. It was responsible for OSPRay
path traced images being washed out. A better solution for handling
sRGB texture images will have to account for all color sources
in the scene. Ensuring that all of them are either linear or scaled
and then applying the framebuffer wide conversion only when all are
scaled. Currently scaled is the exception in visualization so this
is the better behavior.