vtkBlockSortHelper: Attempt to handle disconnected blocks
- Until now, vtkBlockSortHelper did not make any decision when the
blocks are disconnected, yet,  it always reversed the input vector
 and copied that into the result vector.
- As a result, every render caused back blocks to be drawn last and the
blocks in the front would render first and this alternates the next
frame causing a flicker effect.
- Addresses https://gitlab.kitware.com/paraview/paraview/-/issues/22262
- Implements algorithm described in https://gitlab.kitware.com/vtk/vtk/-/issues/19048
- All blocks that are connected are placed inside groups.
- The blocks within those groups are sorted as before. Note that now,
ComputeOrderWithUncertaininty is asked to not allow disconnected bounding
boxes.
- The groups themselves are sorted by computing the overall bounding box
of the group. However, this time, ComputeBoundsWithUncertainity is asked to allow
disconnected bounding boxes.
- There are still some edge cases that are not handled - ex two blocks
share a tiny amount of 3-dimension overlap.