Update vtk cell Triangulate to return only the local cell pt ids
This commit adds two functions to vtkCell interface:
TriangulatePtIds and TriangulateLocalCellPtIds
Note that Triangulate is maintained and not deprecated.
TriangulateLocalCellPtIds is implemented for each cell,
and returns the local point ids.
These points are the used but TriangulatePtIds and Triangulate.
The difference between these two functions is that Triangulate returns
the point and ids, whereas TriangulatePtIds only returns the ids.
The advantage of having a function that does not return the coordinates
is that most of the time, only the ids are required.
This is specifically true for the filters vtkDataSetSurfaceFilter.
For Bezier cells, evaluating these coordinates is really expensive,
so avoiding its computation drastically improves performances.