Slightly speed up vtkDataSetSurfaceFilter for nonlinear cells.
During the triangulation step of vtkDataSetSurfaceFilter,
the local point id is required.
But before, the Triangulate function was only returning the global
 point ids of the cell. So we were iterating over all the global
point ids of a cell until it matches the given global id,
to recover the local id.
In this commit, we directly use TriangulateLocalCellPtIds to get
the local id. No lookup is needed anymore