Improve speed of vtkUnstructuredGridGeometryFilter for higher order cells
In vtkUnstructuredGridGeometryFilter::RequestData, we iterate over
all the cells and grab their point ids.
For most of the cells, the almost non-op function GetFaceArray is called.
However, for higher order cells, the face point ids are not constant
because they depend on the cell order.

What was done before to obtain these ids is that a genericCell was
created (including setting point positions, rational weights, allocating
memory for their approximate linear counterpart, allocating memory for
faces, edges ...).
Then for each face of 3d cells, the point positions, the point ids and
the rational weight were set.

This is really time and memory consuming. In this work, static functions
are instead called to directly obtain these face ids,
without the need to create the cell.