Volume label mapping with mismatched labels in mask and property
The label mapping support in the GPU volume mapper is using texture
lookup of mask values to identify the voxel color and opacity. The
mask value is looked up from a 2D texture generated from the label
transfer functions on the property. If there is a mismatch in the mask
labels and the label transfer functions in the property, the texture
lookup would produce artifacts. This change fixes the lookup value scale
and bias based on the property functions.