ParallelDIY2: Add new data redistribution filter
Add the vtkRedistributeDataSetToSubComm filter, which aggregates data onto
specified mpi ranks for writing to disk or other processing tasks. To
specifiy the ranks where data should be aggregated, you must provide an
mpi sub-communicator configured for only the target mpi processes. The
target processes must be a subset of all processes in the global
communicator.