vtkTableBasedClipDataSet: Allow processing supported and unsupported cells
vtkTableBasedClipDataSet used to process only the cells that it could,
and pass the remaining in vtkClipDataSet. This functionality was
removed when the multithreading of vtkTableBasedClipDataSet took place,
and if any cell could not be processed the whole mesh was passed to vtkClipDataSet.
Now this functionality is added back again in a way simpler and more efficient
way using vtkExtractCells, which is smart enough to just shallow copy the input
if the whole input can't be processed.