Relpace the fixed scalar with a 3D vector for the gravity
vtkLagrangianBasicIntegrationModel implementation using article : "Matida, E. A., et al. "Improved numerical simulation of aerosol deposition in an idealized mouth-throat." Journal of Aerosol Science 35.1 (2004): 1-19."

The code strictly follows the particle equation of motion in this article, which uses a fixed scalar number (G = 9.8) for the z-component of gravity acceleration. This change would like to replace the fixed scalar with a 3D vector as the two reasons below:

1. The direction of gravity may not be the -z direction
2. There gravity can be zero in the simulation