vtkUnstructuredGridGeometryFilter now supports degenerate cells
Degenerate cells are cells for which one edge or face is collapsed.
It means that a single cell contains duplicated ids.
In vtkUnstructuredGridGeometryFilter::InsertFace, the index of the
smallest point cell id was used to positionned a cell relative to
another one. But if the point is not unique, we will have multiple
indices, and we can't use these indices for relative positioning.
So instead of using the smallest point id, we use the smallest
unique point id.